requirements.txt
pygame==1.9.6
numpy==1.18.1
pyyaml==5.3
matplotlib==3.2.0